% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aiq}
\alias{aiq}
\title{The Art Interest Questionnaire data}
\format{
A data frame with 1,269 rows and 4 columns:
\describe{
\item{Group}{Experimental group, factor with three levels: "NonArt-Online", "Art-Online", and "Art-Gallery"}
\item{ID}{Participant ID, factor with 127 levels}
\item{Question}{Factor with ten levels corresponding to individual questions.}
\item{Response}{Response on a 7-point Likert scale}
}
}
\source{
\url{https://osf.io/zf476/}
}
\usage{
aiq
}
\description{
The Art Interest Questionnaire data from the "Art has the power to stretch the boundaries of moral behaviour" manuscript.
}
\keyword{datasets}
