% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{art}
\alias{art}
\title{Art has the power to stretch the boundaries of moral behaviour}
\format{
A data frame with 25,400 rows and 6 columns:
\describe{
\item{Group}{Experimental group, factor with three levels: "NonArt-Online", "Art-Online", and "Art-Gallery"}
\item{ID}{Participant ID}
\item{Response}{Response on a 7-point Likert scale}
\item{PaintingKind}{Factor with two levels: "Moral" or "Immoral"}
\item{PaintingID}{ID of individual paintings}
\item{Scale}{Questionnaire scale, factor with 10 levels}
}
}
\source{
\url{https://osf.io/zf476/}
}
\usage{
art
}
\description{
A subset of data from the "Art has the power to stretch the boundaries of moral behaviour"
manuscript.
}
\keyword{datasets}
