% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_response_counts}
\alias{bootstrap_response_counts}
\title{Compute mean and non-parametric bootstrapped percentile confidence intervals}
\usage{
bootstrap_response_counts(
  df,
  formula,
  ci_variable = "P",
  R = 2000,
  CI = 0.97,
  .progress = TRUE
)
}
\arguments{
\item{df}{Table with data}

\item{formula}{Formula that specifies outcome variable, fixed and random effects.}

\item{ci_variable}{Which variable to use for percentile confidence intervals, either \code{"N"} or \code{"P"} (default).}

\item{R}{Number of samples for bootstrapping, default is \code{2000}.}

\item{CI}{Percentile confidence interval, default is \code{0.97}}

\item{.progress}{Logical, whether to show progress bar during bootstrapping. Default is \code{TRUE}.}
}
\value{
data.frame or tibble with fixed effects columns, as well as N, P, LowerCI, and UpperCI.
}
\description{
Compute mean and non-parametric bootstrapped percentile confidence intervals
}
\examples{
avg_prop <- bootstrap_response_counts(art, Response ~ Group + PaintingKind + Scale, R = 10)

avg_count <- bootstrap_response_counts(art,
                                       Response ~ Group + PaintingKind + Scale,
                                       ci_variable = "N",
                                       R = 10)
}
\seealso{
\code{\link[=count_responses]{count_responses()}}, \code{\link[=lower_ci]{lower_ci()}}, \code{\link[=upper_ci]{upper_ci()}}
}
