% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ci_variable}
\alias{check_ci_variable}
\title{Check validity of variable for CI computation}
\usage{
check_ci_variable(ci_variable)
}
\arguments{
\item{ci_variable}{character}
}
\description{
Check that variable for CI computation is either "P" or "N". Stop otherwise.
}
\examples{
check_ci_variable("P")
}
