% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_that_columns_exist}
\alias{check_that_columns_exist}
\title{Check that all columns are present in a table}
\usage{
check_that_columns_exist(df, column_names)
}
\arguments{
\item{df}{data.frame or tibble}

\item{column_names}{vector with column names}
}
\description{
Check that all columns are present in a table, raises exception if any columns are missing
}
\examples{
data(art)
check_that_columns_exist(art, c("Group", "ID"))
}
