% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_and_posterior.R
\name{compute_bootstrap_and_posterior_predictions}
\alias{compute_bootstrap_and_posterior_predictions}
\title{Compute summary based on both boostrapped samples and of posterior predictive distribution}
\usage{
compute_bootstrap_and_posterior_predictions(
  df,
  formula,
  eta,
  cutpoints,
  ci_variable = "P",
  RBoot = 2000,
  RPost = NULL,
  CI = 0.97,
  .progress = TRUE
)
}
\arguments{
\item{df}{Table with data}

\item{formula}{Formula that specifies outcome variable, fixed and random effects.}

\item{eta}{Matrix of samples_n x nrow(df), each row is a single posterior sample}

\item{cutpoints}{Matrix of samples_n x nlevels(df$outcome) - 1, each row is a single posterior sample}

\item{ci_variable}{Which variable to use for percentile confidence intervals, either \code{"N"} or \code{"P"} (default).}

\item{RBoot}{Number of samples for bootstrapping, default is \code{2000}.}

\item{RPost}{Number of posterior samples to process, \code{NULL} (default) means all samples are used.}

\item{CI}{Percentile confidence interval, default is \code{0.97}}

\item{.progress}{Logical, whether to show progress bar during bootstrapping. Default is \code{TRUE}.}
}
\value{
data.frame or tibble with fixed effects columns, PosteriorMean, PosteriorMedian, PosteriorLowerCI, and PosteriorUpperCI.
}
\description{
Computes bootstrapped summary via \code{\link[=bootstrap_response_counts]{bootstrap_response_counts()}} and posterior predictions via
\code{\link[=posterior_predictions_for_response_counts]{posterior_predictions_for_response_counts()}}, then joins them on fixed factors and response variable
as specified by the formula
}
\examples{
data(aiq)
data(aiq_draws)
eta_draws <- extract_stan_posterior_matrix(aiq_draws, eta)
cutpoints_draws <- extract_stan_posterior_matrix(aiq_draws, cutpoints)
aiq_avg <-
  compute_bootstrap_and_posterior_predictions(aiq,
                                              Response ~ Group + Question,
                                              eta_draws,
                                              cutpoints_draws,
                                              RBoot=10,
                                              RPost=10)
}
\seealso{
\code{\link[=bootstrap_response_counts]{bootstrap_response_counts()}} \code{\link[=posterior_predictions_for_response_counts]{posterior_predictions_for_response_counts()}}
}
