% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stan_posterior_matrix.R
\name{extract_stan_posterior_matrix}
\alias{extract_stan_posterior_matrix}
\title{Extract variable from draws table as a matrix}
\usage{
extract_stan_posterior_matrix(draws_df, varname)
}
\arguments{
\item{draws_df}{Table in \code{draws_df} format}

\item{varname}{Variable name that corresponds to multiple columns in format \code{var[1]}, \code{var[2]}, etc. Can be written directly \code{varname} or as a string \code{"varname"}.}
}
\value{
Matrix
}
\description{
Extract variable from draws table as a matrix
}
\examples{
data(aiq_draws)
cutpoints <- extract_stan_posterior_matrix(aiq_draws, cutpoints)

# same but using string
cutpoints <- extract_stan_posterior_matrix(aiq_draws, "cutpoints")
}
