% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_grid_hlines.R
\name{geom_grid_hlines}
\alias{geom_grid_hlines}
\title{Replicate horizontal grid lines}
\usage{
geom_grid_hlines(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  show.legend = NA,
  inherit.aes = TRUE,
  color = NULL
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{position}{A position adjustment to use on the data for this layer.}

\item{...}{Other arguments passed on to \code{layer()}'s params argument.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.}

\item{color}{Line color, if \code{NULL} theme panel grid line color is used.}
}
\description{
Replicate horizontal grid lines that are hidden by stripes.
}
\examples{
stripes_df <- data.frame(x = c(1, 2, 3), xend = c(2, 3, 4))
ggplot2::ggplot(stripes_df) +
  geom_gradient_vstripe(ggplot2::aes(x = x, xend = xend)) +
  geom_grid_hlines()
}
