% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_vstripe_flip.R
\name{geom_vstripe_flip}
\alias{geom_vstripe_flip}
\title{Vertical stripe with constant fill}
\usage{
geom_vstripe_flip(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  show.legend = NA,
  inherit.aes = FALSE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{position}{A position adjustment to use on the data for this layer.}

\item{...}{Other arguments passed on to \code{layer()}'s params argument.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.}
}
\description{
Displays a vertical stripe that spans the entire height of the plot for each pair of \code{x} and  \code{xend} values. Optional \code{alpha} defaults to \code{1}.
}
\examples{
stripes_df <- data.frame(x = c(1, 2, 3),
                         xend = c(2, 3, 4),
                         fill = c("lightgray", "darkgray", "lightgray"))
ggplot2::ggplot(stripes_df) +
  geom_vstripe_flip(ggplot2::aes(y = x, yend = xend), fill = stripes_df$fill)
}
