% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Inverse logit, wrapper for \code{1 / (1 + exp(-x))}
}
\examples{
inv_logit(-Inf)
inv_logit(0)
inv_logit(Inf)
}
