% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis.R
\name{lower_ci}
\alias{lower_ci}
\title{Compute lower limit for percentile confidence interval}
\usage{
lower_ci(x, CI = 0.97)
}
\arguments{
\item{x}{values}

\item{CI}{confidence interval, defaults to 0.97}
}
\value{
float
}
\description{
Compute lower limit for percentile confidence interval
}
\examples{
# default 97\% CI
lower_ci(rnorm(100))

# alternative CI
lower_ci(rnorm(100), CI=0.89)
}
