% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_logistic.R
\name{ordered_logistic}
\alias{ordered_logistic}
\alias{dordered_logistic}
\alias{rordered_logistic}
\title{Ordered logistic distribution}
\usage{
dordered_logistic(k, eta, c, log = FALSE)

rordered_logistic(n, eta, c)
}
\arguments{
\item{k}{response level}

\item{eta}{vector of means}

\item{c}{vector of cutpoints}

\item{log}{logical, if \code{TRUE} probabilities p are given as \code{log(p)}}

\item{n}{number of observations}
}
\value{
\code{dordered_logistic} gives the density, \code{rordered_logistic} gives random value.
}
\description{
Ordered logistic distribution
}
\examples{
dordered_logistic(2, 0.5, c(-2, -1, 0, 1, 2))
rordered_logistic(1, 0.5, c(-2, -1, 0, 1, 2))
}
