% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse formula into outcome, fixed, and random effects}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{Formula that describes variables}
}
\value{
list with \code{outcome}, \code{fixed}, \code{random} and \code{all} fields
}
\description{
Parse formula in format outcome ~ fixed_effect1 + fixed_effect2 + ... (1 | random_effect1) + (1 | random_effect2) + ...
}
\examples{
# only one fixed effect
parse_formula(y ~ x)

# two fixed effects
parse_formula(y ~ x1 + x2)

# one random effect
parse_formula(y ~ (1|z))

# one fixed and two random effects
parse_formula(y ~ x + (1|z1) + (1|z2))
}
