% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensity.R
\name{plot_intensity}
\alias{plot_intensity}
\title{Plot estimated intensity as box plots}
\usage{
plot_intensity(
  df,
  intensity,
  cutpoints,
  group = NULL,
  fill_interval = c("gray90", "gray85"),
  x_labels_pos = "bottom"
)
}
\arguments{
\item{df}{Table with estimated response intensity}

\item{intensity}{Variable for estimated response intensity}

\item{cutpoints}{Vector with estimated cut points}

\item{group}{Variable for grouping for data or \code{NULL} if there is only a single estimate}

\item{fill_interval}{Fill colors for stripes that indicate intervals.}

\item{x_labels_pos}{Either \code{"bottom"} (default) or \code{"top"}. The latter is useful for combined plots.}
}
\description{
Plot estimated intensity as box plots
}
\examples{
data("aiq")
data("aiq_draws")

# extract estimated intensity
intensity_df <- parse_stan_posterior_array(aiq_draws, a, list("Group" = levels(aiq$Group)))
intensity_df[["Intensity"]] <- inv_logit(intensity_df[["Value"]])

# extract cutpoints
cutpoints <- inv_logit(colMeans(extract_stan_posterior_matrix(aiq_draws, cutpoints)))
plot_intensity(intensity_df, Intensity, cutpoints, group = Group)

# ignoring grouping
plot_intensity(intensity_df, Intensity, cutpoints)
}
