% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictions.R
\name{posterior_predictions_for_response_counts}
\alias{posterior_predictions_for_response_counts}
\title{Compute summary of posterior predictive distribution for the data given posterior distributions of eta and cutpoints}
\usage{
posterior_predictions_for_response_counts(
  df,
  formula,
  eta,
  cutpoints,
  ci_variable = "P",
  R = NULL,
  CI = 0.97,
  .progress = TRUE
)
}
\arguments{
\item{df}{Table with data}

\item{formula}{Formula that specifies outcome variable, fixed and random effects.}

\item{eta}{Matrix of samples_n x nrow(df), each row is a single posterior sample}

\item{cutpoints}{Matrix of samples_n x nlevels(df$outcome) - 1, each row is a single posterior sample}

\item{ci_variable}{Which variable to use for percentile confidence intervals, either \code{"N"} or \code{"P"} (default).}

\item{R}{Number of samples to process, \code{NULL} (default) means all samples are used.}

\item{CI}{Percentile confidence interval, default is \code{0.97}}

\item{.progress}{Logical, whether to show progress bar during bootstrapping. Default is \code{TRUE}.}
}
\value{
data.frame or tibble with fixed effects columns, PosteriorMean, PosteriorMedian, PosteriorLowerCI, and PosteriorUpperCI.
}
\description{
Compute summary of posterior predictive distribution for the data given posterior distributions of eta and cutpoints
}
\examples{
data(aiq)
data(aiq_draws)
eta_draws <- extract_stan_posterior_matrix(aiq_draws, eta)
cutpoints_draws <- extract_stan_posterior_matrix(aiq_draws, cutpoints)
aiq_posterior_avg <-
  posterior_predictions_for_response_counts(aiq,
                                            Response ~ Group + Question,
                                            eta_draws,
                                            cutpoints_draws,
                                            R=10)
}
