% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis.R
\name{upper_ci}
\alias{upper_ci}
\title{Compute upper limit for percentile confidence interval}
\usage{
upper_ci(x, CI = 0.97)
}
\arguments{
\item{x}{values}

\item{CI}{confidence interval, defaults to 0.97}
}
\value{
float
}
\description{
Compute upper limit for percentile confidence interval
}
\examples{
# default 97\% CI
upper_ci(rnorm(100))

# alternative CI
upper_ci(rnorm(100), CI=0.89)
}
